Fixed dependecy cycle error
