+++ scripts/product-api.js
@@ -1,18 +1,28 @@
 // This file contains all the product API endpoints
 
-import { resolveApiPath } from './utils.js';
-
 // prod
- export const API_HOST = 'https://www.ingredion.com';
+export const API_HOST = 'https://www.ingredion.com';
 
 // stage - testing (todo: remove before go-live)
-//export const API_HOST = 'https://ingredion-stage65.adobecqms.net';
+  //export const API_HOST = 'https://ingredion-stage65.adobecqms.net';
 
 export const getUrlParams = () => {
   const params = new URLSearchParams(window.location.search);
   return {
     productName: params.get('name'),
   };
+};
+
+export const resolveApiPath = (apiKey, region, locale) => {
+  const key = `${region}-${locale}`;
+  const override = API_PATH_OVERRIDES[apiKey]?.[key];
+  if (override) return override;
+
+  const defaultBuilder = DEFAULT_PATHS[apiKey];
+  if (defaultBuilder) return defaultBuilder(region, locale);
+
+  console.warn(`No path defined for ${apiKey}`);
+  return '';
 };
 
 export const DEFAULT_PATHS = {
