+++ scripts/utils.js
@@ -1,4 +1,3 @@
-import { DEFAULT_PATHS, API_PATH_OVERRIDES } from './product-api.js';
 
 // valid regions and locales
 const VALID_REGIONS = ['na', 'sa', 'emea', 'apac'];
@@ -149,15 +148,3 @@ export function getCookie(name) {
   }
   return null;
 }
-
-export const resolveApiPath = (apiKey, region, locale) => {
-  const key = `${region}-${locale}`;
-  const override = API_PATH_OVERRIDES[apiKey]?.[key];
-  if (override) return override;
-
-  const defaultBuilder = DEFAULT_PATHS[apiKey];
-  if (defaultBuilder) return defaultBuilder(region, locale);
-
-  console.warn(`No path defined for ${apiKey}`);
-  return '';
-};
